requirements.txt
fastapi
uvicorn
pymongo
jinja2
python-dotenv
